backend/requirements.txt
Django==3.2.25
djangorestframework==3.12.4
django-cors-headers==3.13.0
psycopg2-binary==2.9.9   # было 2.9.3
